Creates a deploy script and establishes an ssh connection
Adds a deploy script to push along our CI/CD pipeline.

Currently, the YAML file should only establish an ssh connection between our repository and our Ec2 instance. We should know that this has worked via a simple echo message, and we should receive a green check mark through GitHub actions. 

This should ONLY activate regarding pushes to main, if and when we want to implement this to some effect with a development branch, we would need an entirely different YAML file to achieve this.

@tabato