Abstract database connection from Flask
Separated the class for the database connection into a separate Python
file called mongoManager.py. To interact with the database, you now
must import the MongoManager class from the file.

Added a flag for the MongoManager initializer to construct a mongomock
connection used for testing, or for an actual connection to interact
with our running database.