Refactor API structure
Renamed mongo_manager to database_manager that contains the
previous MongoManager class. This file will now be in charge of
all database interaction.

Renamed mongo_queries to query_manager that contains the previous
MongoQueries class. This makes a call to the database manager
given when it is initialized in order to query the database.

Changed launch() functionality in Flask app.py. It now requires
the user to specify what database they want to use. 'mongo' is
used for the actual MongoDB containing the data, and 'mockMongo'
is used for testing with the mock database.

Modified Dockerfile that incorporates the new launch()
functionality.